simplify intpt2 driver
